+++ install/install.sh
@@ -2,8 +2,8 @@
 
 WF_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
-EXEC_DIR=`dirname $WF_DIR`
+EXEC_DIR=`dirname "$WF_DIR"`
 
 sudo mkdir -p /usr/local/bin
-sudo rm -rf /usr/local/bin/gitflow
-sudo ln -s $EXEC_DIR/workflow.sh /usr/local/bin/gitflow
+sudo rm -f /usr/local/bin/gitflow
+sudo ln -s "$EXEC_DIR/workflow.sh" /usr/local/bin/gitflow